# contracts/openapi.json
{
  "components": {
    "schemas": {
      "APIKeyCreateForm": {
        "properties": {
          "expires_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires At"
          },
          "name": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name"
        ],
        "title": "APIKeyCreateForm",
        "type": "object"
      },
      "APIKeyCreateResponse": {
        "properties": {
          "api_key": {
            "title": "Api Key",
            "type": "string"
          },
          "key": {
            "$ref": "#/components/schemas/APIKeyResponse"
          }
        },
        "required": [
          "api_key",
          "key"
        ],
        "title": "APIKeyCreateResponse",
        "type": "object"
      },
      "APIKeyErrorDetail": {
        "properties": {
          "details": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Details"
          },
          "error": {
            "enum": [
              "API_KEY_INVALID",
              "API_KEY_USER_MISMATCH",
              "API_KEY_CREATE_FAILED",
              "API_KEY_NAME_ALREADY_EXISTS",
              "API_KEY_NOT_FOUND",
              "API_KEY_UPDATE_FAILED"
            ],
            "example": "API_KEY_NOT_FOUND",
            "title": "Error",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "error",
          "message"
        ],
        "title": "APIKeyErrorDetail",
        "type": "object"
      },
      "APIKeyErrorResponse": {
        "properties": {
          "detail": {
            "$ref": "#/components/schemas/APIKeyErrorDetail"
          }
        },
        "required": [
          "detail"
        ],
        "title": "APIKeyErrorResponse",
        "type": "object"
      },
...
      "APIKeyResponse": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "expires_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires At"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "key_prefix": {
            "title": "Key Prefix",
            "type": "string"
          },
          "last_used_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Used At"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "request_count": {
            "title": "Request Count",
            "type": "integer"
          },
          "revoked_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Revoked At"
          }
        },
        "required": [
          "id",
          "name",
          "key_prefix",
          "created_at",
          "request_count"
        ],
        "title": "APIKeyResponse",
        "type": "object"
      },
...
      "APIKeysResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/APIKeyResponse"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "items"
        ],
        "title": "APIKeysResponse",
        "type": "object"
      },
...
      "AppConfigResponse": {
        "properties": {
          "api_base_path": {
            "title": "Api Base Path",
            "type": "string"
          },
          "app_mode": {
            "enum": [
              "development",
              "production"
            ],
            "title": "App Mode",
            "type": "string"
          },
          "billing_enabled": {
            "title": "Billing Enabled",
            "type": "boolean"
          },
          "bootstrap_access_token": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bootstrap Access Token"
          },
          "bootstrap_user": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/UserResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "email_enabled": {
            "title": "Email Enabled",
            "type": "boolean"
          },
          "frontend_base_path": {
            "title": "Frontend Base Path",
            "type": "string"
          },
          "login_enabled": {
            "title": "Login Enabled",
            "type": "boolean"
          },
          "oauth_enabled": {
            "title": "Oauth Enabled",
            "type": "boolean"
          },
          "oauth_providers": {
            "items": {
              "type": "string"
            },
            "title": "Oauth Providers",
            "type": "array"
          }
        },
        "required": [
          "api_base_path",
          "app_mode",
          "login_enabled",
          "frontend_base_path",
          "email_enabled",
          "oauth_enabled",
          "billing_enabled",
          "oauth_providers"
        ],
        "title": "AppConfigResponse",
        "type": "object"
      },
      "AuthErrorDetail": {
        "properties": {
          "details": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Details"
          },
          "error": {
            "enum": [
              "ACCOUNT_BILLING_REVIEW_REQUIRED",
              "ACCOUNT_DELETE_CODE_INVALID",
              "ACCOUNT_DELETE_CODE_THROTTLED",
              "ACCOUNT_DELETE_CODE_SEND_FAILED",
              "ACCOUNT_DELETE_FAILED",
              "ACCOUNT_DELETE_CONFIRMATION_REQUIRED",
              "LAST_ADMIN_REQUIRED",
              "ADMIN_USERS_FAILED",
              "SIGNUP_FAILED",
              "EMAIL_ALREADY_EXISTS",
              "INVALID_CREDENTIALS",
              "ACCOUNT_LOCKED",
              "EMAIL_NOT_VERIFIED",
              "LOGIN_DISABLED",
              "EMAIL_DISABLED",
              "INVALID_TOKEN",
              "INSUFFICIENT_ROLE",
              "USER_NOT_FOUND",
              "PROFILE_UPDATE_FAILED",
              "OAUTH_PROVIDER_NOT_ENABLED",
              "OAUTH_PROVIDER_CONFIG_INVALID",
              "OAUTH_IDENTITY_CONFLICT",
              "OAUTH_SIGNUP_FAILED",
              "OAUTH_PROVIDER_REQUEST_FAILED"
            ],
            "example": "INVALID_TOKEN",
            "title": "Error",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "error",
          "message"
        ],
        "title": "AuthErrorDetail",
        "type": "object"
      },
      "AuthErrorResponse": {
        "properties": {
          "detail": {
            "$ref": "#/components/schemas/AuthErrorDetail"
          }
        },
        "required": [
          "detail"
        ],
        "title": "AuthErrorResponse",
        "type": "object"
      },
...
      "BillingConfigResponse": {
        "properties": {
          "enabled": {
            "title": "Enabled",
            "type": "boolean"
          },
          "livemode": {
            "title": "Livemode",
            "type": "boolean"
          },
          "publishable_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Publishable Key"
          }
        },
        "required": [
          "enabled",
          "livemode"
        ],
        "title": "BillingConfigResponse",
        "type": "object"
      },
      "BillingErrorDetail": {
        "properties": {
          "details": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Details"
          },
          "error": {
            "enum": [
              "BILLING_WEBHOOK_INVALID",
              "BILLING_WEBHOOK_UNAVAILABLE",
              "BILLING_DISABLED",
              "BILLING_UNAVAILABLE",
              "BILLING_PLAN_UNAVAILABLE",
              "BILLING_CHECKOUT_CONFLICT",
              "BILLING_CHANGE_CONFLICT",
              "BILLING_METHOD_REQUIRED",
              "BILLING_NOT_FOUND",
              "BILLING_RECONCILIATION_REQUIRED"
            ],
            "example": "BILLING_DISABLED",
            "title": "Error",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "error",
          "message"
        ],
        "title": "BillingErrorDetail",
        "type": "object"
      },
      "BillingErrorResponse": {
        "properties": {
          "detail": {
            "$ref": "#/components/schemas/BillingErrorDetail"
          }
        },
        "required": [
          "detail"
        ],
        "title": "BillingErrorResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "InternalErrorResponse": {
        "properties": {
          "error": {
            "const": "INTERNAL_ERROR",
            "title": "Error",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "error",
          "message"
        ],
        "title": "InternalErrorResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "APIKeyHeader": {
        "in": "header",
        "name": "X-API-Key",
        "type": "apiKey"
      },
      "OAuth2PasswordBearer": {
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/auth/token"
          }
        },
        "type": "oauth2"
      }
    }
  },
  "info": {
    "title": "B4 API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/api-keys": {
      "get": {
        "operationId": "list_api_keys_api_v1_api_keys_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIKeysResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/AuthErrorResponse"
                    },
                    {
                      "$ref": "#/components/schemas/APIKeyErrorResponse"
                    }
                  ],
                  "title": "Response 401 List Api Keys Api V1 Api Keys Get"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/AuthErrorResponse"
                    },
                    {
                      "$ref": "#/components/schemas/APIKeyErrorResponse"
                    }
                  ],
                  "title": "Response 403 List Api Keys Api V1 Api Keys Get"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "example": {
                  "detail": {
                    "error": "USER_NOT_FOUND",
                    "message": "User not found."
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/AuthErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InternalErrorResponse"
                }
              }
            },
            "description": "Unexpected server error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          },
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "List Api Keys",
        "tags": [
          "API Keys"
        ]
      },
      "post": {
        "operationId": "create_api_key_api_v1_api_keys_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/APIKeyCreateForm"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIKeyCreateResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/AuthErrorResponse"
                    },
                    {
                      "$ref": "#/components/schemas/APIKeyErrorResponse"
                    }
                  ],
                  "title": "Response 401 Create Api Key Api V1 Api Keys Post"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/AuthErrorResponse"
                    },
                    {
                      "$ref": "#/components/schemas/APIKeyErrorResponse"
                    }
                  ],
                  "title": "Response 403 Create Api Key Api V1 Api Keys Post"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "example": {
                  "detail": {
                    "error": "USER_NOT_FOUND",
                    "message": "User not found."
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/AuthErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "example": {
                  "detail": {
                    "error": "API_KEY_NAME_ALREADY_EXISTS",
                    "message": "API key name already exists."
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/APIKeyErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "500": {
            "content": {
              "application/json": {
                "example": {
                  "detail": {
                    "error": "API_KEY_CREATE_FAILED",
                    "message": "Failed to create API key."
                  }
                },
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/APIKeyErrorResponse"
                    },
                    {
                      "$ref": "#/components/schemas/InternalErrorResponse"
                    }
                  ]
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          },
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Create Api Key",
        "tags": [
          "API Keys"
        ]
      }
    },
...
    "/api/v1/billing/config": {
      "get": {
        "operationId": "billing_config_api_v1_billing_config_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BillingConfigResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "example": {
                  "detail": {
                    "error": "BILLING_WEBHOOK_INVALID",
                    "message": "Invalid billing webhook."
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/BillingErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/AuthErrorResponse"
                    },
                    {
                      "$ref": "#/components/schemas/APIKeyErrorResponse"
                    }
                  ],
                  "title": "Response 401 Billing Config Api V1 Billing Config Get"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/AuthErrorResponse"
                    },
                    {
                      "$ref": "#/components/schemas/APIKeyErrorResponse"
                    }
                  ],
                  "title": "Response 403 Billing Config Api V1 Billing Config Get"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/AuthErrorResponse"
                    },
                    {
                      "$ref": "#/components/schemas/BillingErrorResponse"
                    }
                  ],
                  "title": "Response 404 Billing Config Api V1 Billing Config Get"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "example": {
                  "detail": {
                    "error": "BILLING_CHECKOUT_CONFLICT",
                    "message": "An existing subscription or another checkout requires attention."
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/BillingErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InternalErrorResponse"
                }
              }
            },
            "description": "Unexpected server error"
          },
          "502": {
            "content": {
              "application/json": {
                "example": {
                  "detail": {
                    "error": "BILLING_UNAVAILABLE",
                    "message": "Payment provider is temporarily unavailable."
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/BillingErrorResponse"
                }
              }
            },
            "description": "Bad Gateway"
          },
          "503": {
            "content": {
              "application/json": {
                "example": {
                  "detail": {
                    "error": "BILLING_WEBHOOK_UNAVAILABLE",
                    "message": "Billing webhook is not configured."
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/BillingErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          },
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Billing Config",
        "tags": [
          "Billing"
        ]
      }
    },
...
    "/config": {
      "get": {
        "operationId": "config_config_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppConfigResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InternalErrorResponse"
                }
              }
            },
            "description": "Unexpected server error"
          }
        },
        "summary": "Config"
      }
    },
...
  }
}